+++ test/StoneCutter/CLI/sccomp_nochisel.sh
@@ -1,31 +1,36 @@
 #!/bin/bash
 
-CGCLI_PATH=$1
+SCCOMP_PATH=$1
+
+if [ ! -f $SCCOMP_PATH/sccomp ]; then
+  echo "COULD NOT FIND $SCCOMP_PATH/sccomp"
+  exit -1
+fi
 
 FILE=test.OPT.sc
 touch $FILE
 echo "# this is a stonecutter source file" >> $FILE 2>&1
 
-$CGCLI_PATH/cgcli -D $FILE
+$SCCOMP_PATH/sccomp -D $FILE
 retVal=$?
 if [ ! $? -eq 0 ]; then
-  echo "$CGCLI_PATH/sccomp -D $FILE failed with return code = $retVal"
+  echo "$SCCOMP_PATH/sccomp -D $FILE failed with return code = $retVal"
   rm -Rf $FILE
   exit $retVal
 fi
 
-$CGCLI_PATH/cgcli -disable-chisel $FILE
+$SCCOMP_PATH/sccomp -disable-chisel $FILE
 retVal=$?
 if [ ! $? -eq 0 ]; then
-  echo "$CGCLI_PATH/sccomp -disable-chisel $FILE failed with return code = $retVal"
+  echo "$SCCOMP_PATH/sccomp -disable-chisel $FILE failed with return code = $retVal"
   rm -Rf $FILE
   exit $retVal
 fi
 
-$CGCLI_PATH/cgcli --disable-chisel $FILE
+$SCCOMP_PATH/sccomp --disable-chisel $FILE
 retVal=$?
 if [ ! $? -eq 0 ]; then
-  echo "$CGCLI_PATH/sccomp --disable-chisel $FILE failed with return code = $retVal"
+  echo "$SCCOMP_PATH/sccomp --disable-chisel $FILE failed with return code = $retVal"
   rm -Rf $FILE
   exit $retVal
 fi
