+++ test/StoneCutter/CLI/sccomp_parse.sh
@@ -1,31 +1,36 @@
 #!/bin/bash
 
-CGCLI_PATH=$1
+SCCOMP_PATH=$1
+
+if [ ! -f $SCCOMP_PATH/sccomp ]; then
+  echo "COULD NOT FIND $SCCOMP_PATH/sccomp"
+  exit -1
+fi
 
 FILE=test.PARSE.sc
 touch $FILE
 echo "# this is a stonecutter source file" >> $FILE 2>&1
 
-$CGCLI_PATH/cgcli -p $FILE
+$SCCOMP_PATH/sccomp -p $FILE
 retVal=$?
 if [ ! $? -eq 0 ]; then
-  echo "$CGCLI_PATH/sccomp -p $FILE failed with return code = $retVal"
+  echo "$SCCOMP_PATH/sccomp -p $FILE failed with return code = $retVal"
   rm -Rf $FILE
   exit $retVal
 fi
 
-$CGCLI_PATH/cgcli -parse $FILE
+$SCCOMP_PATH/sccomp -parse $FILE
 retVal=$?
 if [ ! $? -eq 0 ]; then
-  echo "$CGCLI_PATH/sccomp -parse $FILE failed with return code = $retVal"
+  echo "$SCCOMP_PATH/sccomp -parse $FILE failed with return code = $retVal"
   rm -Rf $FILE
   exit $retVal
 fi
 
-$CGCLI_PATH/cgcli --parse $FILE
+$SCCOMP_PATH/sccomp --parse $FILE
 retVal=$?
 if [ ! $? -eq 0 ]; then
-  echo "$CGCLI_PATH/sccomp --parse $FILE failed with return code = $retVal"
+  echo "$SCCOMP_PATH/sccomp --parse $FILE failed with return code = $retVal"
   rm -Rf $FILE
   exit $retVal
 fi
